Discarding changes for libspeex-1.1.1 until it is the version officially used by A1 (Backing out XCode changes)
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1956 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Network/network_microphone_shared.cpp
@@ -152,7 +152,7 @@ int32 copy_and_speex_encode(uint8* outStorage, void* inStorage, int32 inCount, i
     static float storedFrame[160];
     static int storedSamples = 0;
     int bytesWritten = 0;
-    short frame[160];
+    float frame[160];
 	if (inCount + storedSamples < 160) {
 		// not enough to encode a frame; add the data to storedFrame
 	    while (inCount > 0) {
